Refactor RCON
