--- rcon/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "rcon",
-  "version": "1.0.0",
-  "type": "module",
-  "exports": {
-    ".": "./index.js",
-    "./squad": "./squad.js"
-  }
-}
