+++ squad-server/package.json
@@ -13,7 +13,6 @@
     "mysql2": "^2.2.5",
     "pg": "^8.5.1",
     "pg-hstore": "^2.3.3",
-    "rcon": "1.0.0",
     "sequelize": "^6.3.5",
     "sqlite3": "^5.0.0",
     "tedious": "^9.2.1",
